:truck: move files to settings folder
