+++ app/settings/settings.html
@@ -2,7 +2,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="../../style/app.css">
+  <link rel="stylesheet" href="../style/app.css">
   <title>ASTo settings page</title>
 </head>
 <body>
