+++ app/settings/settingsWindow.js
